Drop daemon autostart file
Do not install an autostart desktop file for the daemon any more. It takes a
lot of CPU resources on startup, and is not needed for most day to day
operations. The SSH/GPG agent functionality does not need it at all, and for
the GUIs (seahorse-preferences and the nautilus plugins) a slightly increased
startup cost is a small price to pay for getting a faster desktop startup.

https://bugzilla.gnome.org/show_bug.cgi?id=608012