pgp: Fix setting key trust to "never"
The value of NEVER is less than UNKNOWN in
libcryptui/cryptui.h, so don't use UNKNOWN as the minimum
trust level. Fixes a comparison of a SeahorseValidity type
with a gpgme_validity_t. Also make the documentation accurate,
and make the relative ordering clearer in the code.

Thanks to Pablo Castellano for debugging.
Closes: bgo#609344.