help: Review password-keyring-create help page
Add a seealso to updating passwords to replace the note from the bottom
of the page. Remove document version number from the revision tag (the
document version corresponds to the package version that it is written
for). Shorten title. Use more active language. Try to make the
instructions more consise (fail at one point, add a TODO to fix later).