pkcs11: fix racy build
seahorse-pkcs11.h needs to be built before the files that include it are
compiled. Unfortunately it is generated by valac, and there doesn't seem
to be any simple way to tell this to Automake. But adding the Vala stamp
file to BUILT_SOURCES tells it to compile the Vala before doing anything
else, which is what we need.

https://bugzilla.gnome.org/show_bug.cgi?id=710069