pgp-subkey: Prettify fingerprints in a smarter way
seahorse_pgp_subkey_calc_fingerprint() assumed that it would only take
strings composed of hex digits and nothing else, and it would group
digits in chunks of four.  Now the function can take strings composed
of hex digits and whitespace, and it will format them in chunks of
four digits with a space between them.

https://bugzilla.gnome.org/show_bug.cgi?id=719872