Avoid binding seahorse to the build-time version of gpg
As a user of gpgme, seahorse should prefer gpg2 over gpg, since gpgme
is more fully-functional when it works with gpg2.

Moreover, seahorse should not assume that the version of gpg that it
was built against is the same version of gpg that it is running
against.

GPGME has allowed a NULL value for the file_name parameter for
gpgme_set_engine_info ever since the function appeared.  This will use
gpgme's default engine selection, which is likely to be better than
hardcoding the path that seahorse found during compile time.

Moreover, seahorse should not bother trying to build against archaic
versions of these branches, and certainly shouldn't hardcode numeric
values that only worked for old versions in the seahorse headers.

This changeset adjusts the configure.ac tests to make sure that the
build environment has a non-archaic version of gnupg at least.

Signed-off-by: Stef Walter <stefw@redhat.com>
 * Remove support for GnuPG 1.4.x as discussed at linked bug

https://bugzilla.gnome.org/show_bug.cgi?id=750468