gkr: Don’t update key description if it doesn’t change
If the user right-clicks in the ‘Description’ entry in the item
properties window, without changing the description, their text
selection will be dropped. This is because Seahorse is trying to update
the item description in the background, as a result of the entry losing
focus.

Avoid trying to update the item description if it has not changed, which
will cover most cases of the user right-clicking.

https://bugzilla.gnome.org/show_bug.cgi?id=777802