Port /src to Vala.
This also brings Application and SearchProvider to /src, where they
belong better, as we can now specify common/libseahorse as a properly
separate library. This was not possible in the previous setup.

Note that the search provider also went through a bit of a change, due
to the abstractions that Vala provides, which gives us this nicer to
follow workflow.