GenerateSelect: use GtkTemplate.
I also added a translatable string for the window title. Apparently, we
didn't even set one before, which looked really bad (especially with the
buttons in the headerbar). I know this technically isn't allowed with
the code freeze, but I think we can agree that an untranslated title is
better than no title at all.