KeyManagerStore: use Gcr.FilterCollection.
In the long term, we should get rid of Seahorse.Collection by replacing
it with Gcr.FilterCollection in other parts of the code base as well,
but I don't want to change too much code so close to release.

Fixes #175, as I can no longer reproduce it with these changes.

https://gitlab.gnome.org/GNOME/seahorse/issues/175