Port most uses of Gtk.Action to GLib.Action
I wasn't planning on doing this in such a huge rework, but given that
the way the Gtk.ActionGroups were handled was so tightly coupled
throughout the codebase, I barely had a choice.

A list of the most major changed:
* Since it was easier to setup (compared to migrating the menubar), we
now use a popover for our menu (together with a GMenuModel in the
builder file).
* We no longer use the registry as a magical way of passing generators.
We make a menu with the GActions that can create keys, and they will be
only shown if that action exists. This means the Registry no longer has
a concept of generators.
* Given that it was the easiest way, we removed the GenerateSelect
dialog for selecting new items in favor of a popover with the actions
set.
* A SeahorseBackend's `actions`-property can no longer be null
* Moved some GActions from KeyManager/Catalog to SeahorseApplication
where appropriate.
* Moved the actions that were being registered from the Generate dialogs
(e.g. Seahorse.Ssh.Generate) to their respective backends.