Port most uses of Gtk.Action to GLib.Action
I wasn't planning on doing this in such a huge rework, but given that
the way the Gtk.ActionGroups were handled was so tightly coupled
throughout the codebase, I barely had a choice.

A list of the most major changed:
* Since it was easier to setup (compared to migrating the menubar), we
now use a popover for our menu (together with a GMenuModel in the
builder file).
* A SeahorseBackend's `actions`-property can no longer be null
* Moved some GActions from KeyManager/Catalog to SeahorseApplication
where appropriate.
* Moved the actions that were being registered from the Generate dialogs
(e.g. Seahorse.Ssh.Generate) to their respective backends.