pgp: KeyserverResults: don't subclass Catalog
We didn't really use the methods from `Seahorse.Catalog` anyway, and we
can now use different UI widgets. Removing this as the parent class
allows us to clean up both Catalog and KeyManager, which are quite
intertwined.

As we're doing some cleaning up anyway, we also cleaned up the UI a bit.