gkr: Backend: load_collections can have errors
Make sure we also call the appropriate finish method of
`SecretService::load_collections()`, as that at least tells us if
something went wrong.

Found out this snipped, as somehow the `Items` property showed an
item that didn't exist as an object path. This made the loading of the
whole collection fail, so I couldn't view the login keyring anymore in
Seahorse, which is quite unfortunate.