pgp: KeyProperties: don't subclass ObjectWidget
SeahorseObjectWidget is really obsolete, as we now have more modern API
like GtkBuilder and GtkTemplate (which we sadly can't port to in a
trivial way).

The purpose of this commit is also partially to make clear the
difference between the public key properties and the private key
properties in an easier way, and to get rid of redundant or obsolete
code.

Finally, we also ported it to use GAction, which means we can simplify
and decouple the UI definition from the underlying code a bit.