Place: Move away from GtkActionGroup
GtkActionGroup was deprecated in favor of GAction, so use that instead.
To provide the necessary information for the UI (such as translatable
strings), we also provide a property of type `GMenuModel`, so that we
don't have to care which acions are available on the SeahorsePlace or
not.

This also solves a regression where we were no longer showing a context
menu in the sidebar if actions were available.