build: Fix build with check-compatible-gpg=false
Before this commit, we only checked the GPG (and GPGME) versions when
the option `check-compatible-gpg` was set. However, if the user wants to
bypass that check (by setting the option to false), we still need to
have the version printed back to us, since we use it later on to set
some variables, like `GPG_MAJOR` etc.

Fixes https://gitlab.gnome.org/GNOME/seahorse/issues/217