pgp: Ellipsize if there are too many uids
The reason for this commit is that some keys have a enormous amount of
UIDs associated with them (think company-wide keys). Thiis makes their
entry in the general list -which shows all UIDs- gets really bloated.

This constrains the amount of showed UIDs to 5. If the user wants to
see more, they will have to go to the key properties dialog.