sidebar: Don't show System Trust when empty
In most cases, 'Default Trust' is empty, as people generally don't touch
it. See the linked issue for the difference between 'System Trust' and
'Default Trust'.

To implement this, we add a "show-if-empty" property to `SeahorsePlace`,
so we can extend this to other types of collections if needed. That way
we can filter empty collections from the sidebar when they don't make
sense and/or confuse users.

Fixes https://gitlab.gnome.org/GNOME/seahorse/issues/252