pgp: Allow exporting both public & secret key
The UI made it a bit unclear, but Seahorse can export both the public
and the private key of a GPG key. The former happened by clicking the
"Export..." button when right-clicking it in the key list, the latter
had its own dedicated button in the key properties dialog.

This commit changes the "Export secret key" button to show a popover
instead, where the user can choose between exporting their secret and
their public key, which makes all of this a bit more intuitive.

Fixes https://gitlab.gnome.org/GNOME/seahorse/issues/208