sidebar: Properly invalidate headers
On each time we called `update_places()`, we added each place in a
sorted fashion, and then _sorted it again_. This was done because we
were otherwise showing double headers. Even more, the internal
implementation of `ListStore.sort()` emits a signal that all elements
have been removed and re-added, which meant that the Sidebar cleared its
selection, triggering some other unwanted and unnecessary callbacks as
well.

In any case, this can be easily solved by calling
`Gtk.ListBox.invalidate_headers()`, which is exactly meant for this kind
of use case.