sidebar: Fix loading items of unlocked places
When clicking the unlock button in the sidebar, we automatically select
that row. In turn, that activates the base collection to start loading
the items that are part of the GkrKeyring, before it's been fully
unlocked. As a consequence, we can't load the labels or other relevant
information yet.

This commit introduces a (temporary) workaround, which clears the
selection (clearing the collection), before selecting the row again,
leading to all the items being loaded _after_ the keyring has been
unlocked.

Fixes https://gitlab.gnome.org/GNOME/seahorse/-/issues/306