pgp: ServerSource: Only use a URI property
We have a redundant property for the (unresolved) address and the URI
that is put in by the user (which is separated by some quick-n-dirty
function to extract the scheme and host). Since we're already using
better URI wrappers inside subclasses (either `SoupUri` or
`GNetworkAddress`), we can just get rid of the "key-server" property,
which didn't amount to much really.

This commits also no longer requires libsoup for resolving addresses in
LDAP sources.