pgp: Make PgpBackend responsible for handling keyservers
That way, we don't need to fiddle around with the `GSettings` keys (and
its special values format) in several places, and we can just expose
what we need. With a bit of luck, it allows us to also get rid of the
`SeahorsePgpSettings` singleton which is strewn around in several
places.