key-manager: Make sure to update after unlocking
When we unlock a keyring, the main list view gets updated due to a
signal that gets triggered when the number of elements in a collection
changes. In the case of an empty keying however, we won't get this
signal as the number of elements will remain zero. We can fix this by
explicitly triggering an update.

Fixes https://gitlab.gnome.org/GNOME/seahorse/-/issues/315