pgp: Expose a key's fields using GListModel
The current API mostly used `GList` (which is a linked list). This
worked well enough for most use cases, but becomes a bit cumbersome to
use with modern GTK APIs, which favor the adaptive `GListModel`. As
such, this commit now exposes the keys, subkeys and UIDs of a PGP key
using a `GListModel` instead of a `GList`.

This also required some internal changes for the `SeahorseGpgmeKey`
subclass, which kept a separate copy of UIDs (and had a lot of back and
forth of vfunc calling between itself and its parent class). The new
class is much simpler and doesn't seem to give any regressions.