pgp-backend: Use correct settings for last-search-servers
This fixes a regression from commit bc26f2bf, which accidentally used
the PGP settings to retrieve the latest search query, rather than
seahorse's own settings.

This commit also cleans up the API for the property a bit, so that it
can be used with a `length` out-parameter from C. While we're at it,
this also fixes a tiny memory leak in the keyserver search dialog.