Create SeahorseDatePicker
Use it to get rid of `EggDateTime`, and the sometimes troublesome parts
of its implementation. `SeahorseDatePicker` is implemented in such a way
that it can be embedded in GtkBuilder .xml files, so we can define it in
a declarative manner, and also don't have to deal with placeholder
widgets.

Fixes https://gitlab.gnome.org/GNOME/seahorse/-/issues/282