pgp: Add a test for HKP
With the advent of HKPS support, some of the code in `SeahorseHkpSource`
got changed quite a lot. With no proper unit tests, it's hard to
guarantee we didn't break anything. With help of the previous commit, we
can now finally start writing tests per backend, so let's do that.

The test actually uncovered some issues with the current HKP
implementetion as well, notably that the first key is ignored. Fix those
while we're at it.