search-provider: Don't escape result description as markup
Only the shell can know how the description string will be used and if
or when it should be escaped. Previously the shell did not escape the
description before displaying it with markup to highlight search hits,
but now it does. For seahorse this however means without this change the
string will be escaped twice and could cause markup to show up in the
description text.

Related: https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/2033