Add seahorse keyword to desktop file
When seahorse is installed on the host system, you can search for "seahorse" in the GNOME shell overview and the app will appear due to the match on the Exec= line of the desktop file. But when seahorse is installed via flatpak, the Exec= line gets rewritten and there is no keyword match. Adding the app name to the list of keywords will help those of us who rely on old muscle memory to launch the app.