gkr: don't show an empty label on network passwords
If the app uses the network password scheme and doesn't set a label,
we might get into an awkward situation where we set the details, but not
the label (while the latter is searchable, the former isn't).

In that case, set the server URL as the label instead of the details.

Related: https://gitlab.gnome.org/GNOME/seahorse/-/issues/388