build: Fix config.h target
Config.h is intended to be generated using vcs_tag, and, per the Meson
reference manual, the return value of vcs_tag must be set as a dependency for
targets that rely on it, or Meson will not know the order in which the
targets need to be built. Also remove the old config_file target so that we
have only one target that generates config.h.

Closes #394