pgp: Move KeyserverResults to KeyserverSearch dialog
Rather than us having a separate dialog for the results of key search,
we can just use `AdwNavigationView` with 2 pages: first the initial
search, and then a seoncd one that pops up to show the search results.

This is also quite a bit nicer from a UX point of view, as we no longer
have several dialogs flashing when all of this is happening.