pgp: Separate logic for key creation parameters
This commit creates a separate object to track all parameters involved
with key parameter creation. On one hand, that allows us to clean up
that logic which was spread over multiple places, and also allows for
easier extension in the future. It could also allow us to store some of
the validation logic later on.