build-aux: Refactor gpg_check.py
Refactor `gpg_check_version.py` into a more generic `gpg_check.py` that
can take multiple commands, of which the only supported one is
"version", which does the exact same thing as before.

This allows us to parse (and use) more of the information from the
output of `gpg --version` later.